Streamline Python entrypoint into __main__
